v 0.3
This version looks like the final one, but it is not. Actually all the needed classes
and methods are written.

For now there are some failures during running related to Unit test of XmlTransfornerTest class.

Maybe there are other bugs, but i think i wikk fix them in the next commit.